Added basic version of the Pause Menu
All buttons function and the pause_game input can be used to leave the settings menu and then leave the pause menu. I had to make a slight change to the Load.gd script because it was not setting the get_tree().current_scene to the new scene instance and it was returning null after the first scene transition. I also removed the deffered call to add_child and just called add_child instead because it was not playing nice with setting get_tree().current_scene.

We need a PAUSED icon to replace the small label.

Thanks =)